--- a/extension/parquet/include/parquet_bss_encoder.hpp
+++ b/extension/parquet/include/parquet_bss_encoder.hpp
@@ -15,8 +15,10 @@ namespace duckdb {
 class BssEncoder {
 public:
 	explicit BssEncoder(const idx_t total_value_count_p, const idx_t bit_width_p)
-	    : total_value_count(total_value_count_p), bit_width(bit_width_p), count(0),
-	      buffer(Allocator::DefaultAllocator().Allocate(total_value_count * bit_width)) {
+	    : total_value_count(total_value_count_p), bit_width(bit_width_p), count(0) {
+		if (total_value_count * bit_width != 0) {
+			buffer = Allocator::DefaultAllocator().Allocate(total_value_count * bit_width);
+		}
 	}
 
 public: