--- a/extension/parquet/include/parquet_bss_encoder.hpp
+++ b/extension/parquet/include/parquet_bss_encoder.hpp
@@ -16,7 +16,7 @@ class BssEncoder {
 public:
 	explicit BssEncoder(const idx_t total_value_count_p, const idx_t bit_width_p)
 	    : total_value_count(total_value_count_p), bit_width(bit_width_p), count(0),
-	      buffer(Allocator::DefaultAllocator().Allocate(MaxValue<idx_t>(total_value_count * bit_width, 1))) {
+	      buffer(Allocator::DefaultAllocator().Allocate(total_value_count * bit_width + 1)) {
 	}
 
 public:
@@ -31,7 +31,7 @@ public:
 
 	void FinishWrite(WriteStream &writer) {
 		D_ASSERT(count == total_value_count);
-		writer.WriteData(buffer.get(), buffer.GetSize());
+		writer.WriteData(buffer.get(), total_value_count * bit_width);
 	}
 
 private: