--- a/test/parquet/bss_roundtrip.test_slow
+++ b/test/parquet/bss_roundtrip.test_slow
@@ -7,7 +7,7 @@ require parquet
 require tpch
 
 statement ok
-call dbgen(sf=0.1);
+call dbgen(sf=0.01);
 
 statement ok
 create view doubles as