@carlopi 2025-01-10
(we were chatting with @Maxxen how the copy constructor here is special) +1 on https://github.com/duckdb/duckdb/pull/15653/commits/8f6f9456777b92d08574611678fd0be431d33877